fix: improve sanity service paths, type names, remove hoc for iconPicker
