+++ src/app/[locale]/(unauth)/curriculo/[slug]/page.tsx
@@ -1,18 +1,17 @@
 import { PortableText } from '@portabletext/react';
 import Image from 'next/image';
-import { getAllTeamMembersInfo, getTeamMemberInfo } from 'sanityClient';
 
 import Button from '@/components/Button';
 import Card from '@/components/Card';
 import DropDownImage from '@/components/DropDownImage/DropDownImage';
-import IconPickerClient from '@/components/IconPickerClient/IconPickerClient';
 import Text from '@/components/Text';
 import WppButton from '@/components/WppButton';
 import CVData from '@/data/cv.json';
 import BrasilBrand from '@/public/assets/images/brands/brasil-brand.png';
+import { getAllTeamMembersInfo, getTeamMemberInfo } from '@/services/sanity/TeamMemberServices';
 // import EUABrand from '@/public/assets/images/brands/eua-brand.png'; TODO: Translate all text to English and use DropDownImage to change.
 import type { UserProfile } from '@/types/cv';
-import type { ITeamMember } from '@/types/sanity';
+import type { TeamMember } from '@/types/sanity';
 
 type Params = {
   slug: string;
@@ -26,7 +25,7 @@ export async function generateStaticParams() {
 }
 
 const CVPage = async ({ params }: { params: Params }) => {
-  const teamMember: ITeamMember = await getTeamMemberInfo(params.slug);
+  const teamMember: TeamMember = await getTeamMemberInfo(params.slug);
   const selectedCV = CVData.find((item: UserProfile) => item.slug === params.slug);
 
   if (!teamMember || !selectedCV) {
@@ -50,7 +49,7 @@ const CVPage = async ({ params }: { params: Params }) => {
           style="outlined-gray"
           icon={(
             <svg width="11" height="10" viewBox="0 0 11 10" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <path fill-rule="evenodd" clip-rule="evenodd" d="M5.5 0C5.76036 0 5.97143 0.197521 5.97143 0.441176V5.84667L7.36665 4.54098C7.55075 4.36869 7.84925 4.36869 8.03335 4.54098C8.21745 4.71327 8.21745 4.99261 8.03335 5.1649L5.83335 7.22372C5.64925 7.39601 5.35075 7.39601 5.16665 7.22372L2.96665 5.1649C2.78255 4.99261 2.78255 4.71327 2.96665 4.54098C3.15075 4.36869 3.44925 4.36869 3.63335 4.54098L5.02857 5.84667V0.441176C5.02857 0.197521 5.23964 0 5.5 0ZM10.5286 6.76471C10.7889 6.76471 11 6.96223 11 7.20588V7.79412C11 8.37915 10.7517 8.94023 10.3096 9.35391C9.86756 9.7676 9.26801 10 8.64286 10H2.35714C1.73199 10 1.13244 9.7676 0.690391 9.35391C0.248341 8.94023 0 8.37915 0 7.79412V7.20824C0 6.96458 0.211066 6.76706 0.471429 6.76706C0.731791 6.76706 0.942857 6.96458 0.942857 7.20824V7.79412C0.942857 8.14514 1.09186 8.48178 1.35709 8.72999C1.62232 8.9782 1.98205 9.11765 2.35714 9.11765H8.64286C9.01795 9.11765 9.37768 8.9782 9.64291 8.72999C9.90814 8.48178 10.0571 8.14514 10.0571 7.79412V7.20588C10.0571 6.96223 10.2682 6.76471 10.5286 6.76471Z" fill="white" />
+              <path fillRule="evenodd" clipRule="evenodd" d="M5.5 0C5.76036 0 5.97143 0.197521 5.97143 0.441176V5.84667L7.36665 4.54098C7.55075 4.36869 7.84925 4.36869 8.03335 4.54098C8.21745 4.71327 8.21745 4.99261 8.03335 5.1649L5.83335 7.22372C5.64925 7.39601 5.35075 7.39601 5.16665 7.22372L2.96665 5.1649C2.78255 4.99261 2.78255 4.71327 2.96665 4.54098C3.15075 4.36869 3.44925 4.36869 3.63335 4.54098L5.02857 5.84667V0.441176C5.02857 0.197521 5.23964 0 5.5 0ZM10.5286 6.76471C10.7889 6.76471 11 6.96223 11 7.20588V7.79412C11 8.37915 10.7517 8.94023 10.3096 9.35391C9.86756 9.7676 9.26801 10 8.64286 10H2.35714C1.73199 10 1.13244 9.7676 0.690391 9.35391C0.248341 8.94023 0 8.37915 0 7.79412V7.20824C0 6.96458 0.211066 6.76706 0.471429 6.76706C0.731791 6.76706 0.942857 6.96458 0.942857 7.20824V7.79412C0.942857 8.14514 1.09186 8.48178 1.35709 8.72999C1.62232 8.9782 1.98205 9.11765 2.35714 9.11765H8.64286C9.01795 9.11765 9.37768 8.9782 9.64291 8.72999C9.90814 8.48178 10.0571 8.14514 10.0571 7.79412V7.20588C10.0571 6.96223 10.2682 6.76471 10.5286 6.76471Z" fill="white" />
             </svg>
           )}
         >
@@ -96,7 +95,7 @@ const CVPage = async ({ params }: { params: Params }) => {
                 size="md"
                 icon={(
                   <svg width="8" height="6" viewBox="0 0 8 6" fill="none" xmlns="http://www.w3.org/2000/svg">
-                    <path fill-rule="evenodd" clip-rule="evenodd" d="M4.51203 4.83217C4.51191 4.83204 4.51171 4.83205 4.51159 4.83217C4.38844 4.96748 4.22157 5.04348 4.04758 5.04348C3.87349 5.04348 3.70652 4.96739 3.58335 4.83193L1.1054 2.1044C0.982224 1.9687 0.913045 1.78468 0.913086 1.59281C0.913127 1.40095 0.982384 1.21697 1.10562 1.08133C1.22886 0.945696 1.39598 0.869523 1.57023 0.869568C1.74447 0.869613 1.91156 0.945874 2.03474 1.08157C3.11368 2.26961 4.98148 2.26961 6.06042 1.08157C6.18426 0.949742 6.35017 0.876744 6.52242 0.878302C6.69467 0.87986 6.85947 0.955849 6.98133 1.0899C7.10318 1.22396 7.17235 1.40535 7.17393 1.59501C7.17551 1.78467 7.10937 1.96743 6.98976 2.10392L4.51247 4.83217C4.51235 4.8323 4.51215 4.8323 4.51203 4.83217Z" fill="white" />
+                    <path fillRule="evenodd" clipRule="evenodd" d="M4.51203 4.83217C4.51191 4.83204 4.51171 4.83205 4.51159 4.83217C4.38844 4.96748 4.22157 5.04348 4.04758 5.04348C3.87349 5.04348 3.70652 4.96739 3.58335 4.83193L1.1054 2.1044C0.982224 1.9687 0.913045 1.78468 0.913086 1.59281C0.913127 1.40095 0.982384 1.21697 1.10562 1.08133C1.22886 0.945696 1.39598 0.869523 1.57023 0.869568C1.74447 0.869613 1.91156 0.945874 2.03474 1.08157C3.11368 2.26961 4.98148 2.26961 6.06042 1.08157C6.18426 0.949742 6.35017 0.876744 6.52242 0.878302C6.69467 0.87986 6.85947 0.955849 6.98133 1.0899C7.10318 1.22396 7.17235 1.40535 7.17393 1.59501C7.17551 1.78467 7.10937 1.96743 6.98976 2.10392L4.51247 4.83217C4.51235 4.8323 4.51215 4.8323 4.51203 4.83217Z" fill="white" />
                   </svg>
                 )}
               >
@@ -120,7 +119,7 @@ const CVPage = async ({ params }: { params: Params }) => {
           <div className="flex gap-2 overflow-x-auto scroll-smooth pb-4 pl-8 lg:container xxs:ml-fluid-xxs xs:ml-fluid-xs sm:ml-fluid-sm mdlg:ml-fluid-mdlg lg:ml-auto lg:flex-wrap lg:gap-4 lg:overflow-hidden lg:pl-8">
             {teamMember.skills.map((skill, index) => (
               <Card key={index} style="outlined-gray" className="flex min-w-40 flex-1 flex-col justify-start space-y-2 pt-20">
-                {skill.icon.name && <IconPickerClient iconName={skill.icon.name} />}
+                {skill.icon.svg && <svg dangerouslySetInnerHTML={{ __html: skill.icon.svg }} className="size-12 text-white" viewBox="0 0 20 20" />}
                 <Text as="p">{skill.name}</Text>
                 <Text as="p" size="xxs" className=" text-secondaryLighter">{skill.description}</Text>
               </Card>
