+++ src/components/CheckBox/CheckBox.tsx
@@ -66,7 +66,7 @@ const CheckBox: React.FC<CheckBoxProps> = ({
           ? (<img src={icon} sizes={iconSize} alt="icon"></img>)
           : (
               <svg width={iconSize} height={iconSize} viewBox="0 0 8 6" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <path fill-rule="evenodd" clip-rule="evenodd" d="M7.37181 0.152764C7.45389 0.250608 7.5 0.383295 7.5 0.521647C7.5 0.659999 7.45389 0.792686 7.37181 0.89053L3.21231 5.84724C3.1302 5.94505 3.01885 6 2.90275 6C2.78665 6 2.6753 5.94505 2.5932 5.84724L0.622908 3.49932C0.543151 3.40092 0.499019 3.26912 0.500017 3.13232C0.501014 2.99551 0.547062 2.86465 0.628241 2.76791C0.709421 2.67117 0.819237 2.6163 0.934038 2.61511C1.04884 2.61392 1.15944 2.66651 1.24202 2.76156L2.90275 4.74059L6.7527 0.152764C6.8348 0.0549493 6.94615 0 7.06225 0C7.17835 0 7.2897 0.0549493 7.37181 0.152764Z" fill="#F8FAFC" />
+                <path fill-rule="evenodd" clipRule="evenodd" d="M7.37181 0.152764C7.45389 0.250608 7.5 0.383295 7.5 0.521647C7.5 0.659999 7.45389 0.792686 7.37181 0.89053L3.21231 5.84724C3.1302 5.94505 3.01885 6 2.90275 6C2.78665 6 2.6753 5.94505 2.5932 5.84724L0.622908 3.49932C0.543151 3.40092 0.499019 3.26912 0.500017 3.13232C0.501014 2.99551 0.547062 2.86465 0.628241 2.76791C0.709421 2.67117 0.819237 2.6163 0.934038 2.61511C1.04884 2.61392 1.15944 2.66651 1.24202 2.76156L2.90275 4.74059L6.7527 0.152764C6.8348 0.0549493 6.94615 0 7.06225 0C7.17835 0 7.2897 0.0549493 7.37181 0.152764Z" fill="#F8FAFC" />
               </svg>
             )}
