+++ src/components/WppButton/WppButton.tsx
@@ -1,6 +1,6 @@
 'use client';
 import Button from '@/components/Button';
-import { redirectToWpp } from '@/utils/redirectToWpp';
+import { redirectToWhatsApp } from '@/utils/redirectToWhatsApp';
 
 type WppButtonProps = Partial<React.ComponentProps<typeof Button>> & {
   cellPhone: string | number;
@@ -17,7 +17,7 @@ const WppButton: React.FC<WppButtonProps> = ({ cellPhone, message, label, classN
       iconRight
       size={size}
       className={className}
-      onClick={() => redirectToWpp(cellPhone, message)}
+      onClick={() => redirectToWhatsApp(cellPhone, message)}
       icon={(
         <svg width="14" height="14" viewBox="0 0 14 14" fill="none" xmlns="http://www.w3.org/2000/svg">
           <path d="M11.5666 2.61373C10.9897 2.03101 10.3026 1.56899 9.54528 1.25462C8.78797 0.940243 7.97564 0.779804 7.15568 0.782661C3.72006 0.782661 0.919968 3.58275 0.919968 7.01837C0.919968 7.90034 1.10565 8.76213 1.45762 9.55523C1.62594 9.9345 1.68927 10.3581 1.58023 10.7584C1.26429 11.9184 2.32205 12.9856 3.48475 12.6799L3.5567 12.661C3.9526 12.557 4.36982 12.6166 4.74733 12.7749C5.50655 13.0933 6.32515 13.2604 7.15568 13.2604C10.5913 13.2604 13.3914 10.4603 13.3914 7.02466C13.3914 5.35719 12.7433 3.7904 11.5666 2.61373ZM7.15568 12.2033C6.22441 12.2033 5.31202 11.9516 4.51289 11.4796C4.39218 11.4072 4.24736 11.3865 4.11121 11.4223L3.88576 11.4816C3.20803 11.6597 2.59161 11.0374 2.77618 10.3614L2.82204 10.1934C2.86095 10.0509 2.83742 9.89855 2.75733 9.77441C2.23994 8.94821 1.96521 7.9932 1.9645 7.01837C1.9645 4.16165 4.29266 1.83348 7.14938 1.83348C8.5337 1.83348 9.83621 2.37462 10.8115 3.35623C11.2945 3.83694 11.6772 4.40872 11.9375 5.03844C12.1978 5.66815 12.3305 6.34327 12.328 7.02466C12.3406 9.88138 10.0124 12.2033 7.15568 12.2033ZM9.99981 8.32718C9.8425 8.25167 9.07484 7.87413 8.93641 7.8175C8.79168 7.76716 8.69101 7.74199 8.58404 7.893C8.47707 8.05031 8.18133 8.40268 8.09324 8.50336C8.00514 8.61033 7.91076 8.62292 7.75345 8.54111C7.59614 8.46561 7.09275 8.29571 6.50127 7.76716C6.03564 7.35186 5.72732 6.84218 5.63293 6.68487C5.54484 6.52757 5.62035 6.44577 5.70215 6.36397C5.77136 6.29475 5.85945 6.18149 5.93496 6.09339C6.01047 6.0053 6.04193 5.93609 6.09227 5.83541C6.14261 5.72844 6.11744 5.64035 6.07969 5.56484C6.04193 5.48933 5.72731 4.72167 5.60147 4.40705C5.47562 4.10502 5.34348 4.14277 5.2491 4.13648H4.94706C4.84009 4.13648 4.67649 4.17423 4.53177 4.33154C4.39334 4.48885 3.99063 4.86639 3.99063 5.63405C3.99063 6.40172 4.55065 7.14422 4.62616 7.24489C4.70166 7.35186 5.72732 8.92495 7.28782 9.59823C7.65906 9.76183 7.94851 9.85621 8.17503 9.92543C8.54628 10.045 8.88607 10.0261 9.15664 9.98835C9.45867 9.94431 10.0816 9.61081 10.2075 9.24586C10.3396 8.8809 10.3396 8.57258 10.2956 8.50336C10.2515 8.43415 10.1571 8.40268 9.99981 8.32718Z" fill="white" />
